+++ Adapters/BidonAdapterBidMachine/Providers/Bidding/BidMachineBiddingDemandProvider.swift
@@ -17,11 +17,23 @@ where AdObject: BidMachineAdProtocol {
     }
 
     func collectBiddingToken(
+        auctionKey: String?,
         biddingTokenExtras: BidMachineAdUnitExtras,
         response: @escaping (Result<String, MediationError>) -> ()
     ) {
-
-        BidMachineSdk.shared.token(with: placementFormat) { token in
+        let key = auctionKey.flatMap { $0.isEmpty ? nil : $0 } ?? "default"
+        let placementId = biddingTokenExtras.placements?[key] as? String
+        let placement = try? BidMachineSdk.shared.placement(from: placementFormat) {
+            if let placementId {
+                $0.withPlacementId(placementId)
+            }
+            $0.withCustomParameters([String: Any]())
+        }
+        guard let placement else {
+            response(.failure(.unspecifiedException("No placement to generate bidding token")))
+            return
+        }
+        BidMachineSdk.shared.token(placement: placement) { token in
             guard let token else {
                 response(.failure(.unspecifiedException("BidMachine has not provided bidding token")))
                 return
@@ -30,39 +42,15 @@ where AdObject: BidMachineAdProtocol {
         }
     }
 
+    func collectBiddingToken(biddingTokenExtras: BidMachineAdUnitExtras, response: @escaping (Result<String, MediationError>) -> ()) {
+
+    }
+
     func load(
         payload: BidMachineBiddingPayload,
         adUnitExtras: BidMachineAdUnitExtras,
         response: @escaping DemandProviderResponse
     ) {
-        do {
-            let configuration = try BidMachineSdk.shared.requestConfiguration(placementFormat)
-            var parameters = adUnitExtras.customParameters ?? [String: String]()
-            parameters["mediation_mode"] = "bidon"
-
-            configuration.populate { builder in
-                builder.withPayload(payload.payload)
-                builder.withCustomParameters(parameters)
-            }
-
-            BidMachineSdk.shared.ad(AdObject.self, configuration) { [weak self] ad, error in
-                guard let self = self else { return }
-
-                guard let ad = ad, error == nil else {
-                    response(.failure(.noBid(error?.localizedDescription)))
-                    return
-                }
-
-                ad.controller = UIApplication.shared.bd.topViewcontroller
-                ad.delegate = self
-
-                self.response = response
-                self.ad = ad
-
-                ad.loadAd()
-            }
-        } catch {
-            response(.failure(.incorrectAdUnitId))
-        }
+        fatalError("BidMachineBiddingDemandProvider is not able to create ad object")
     }
 }