+++ Adapters/BidonAdapterVungle/Providers/VungleAdViewDemandProvider.swift
@@ -11,117 +11,118 @@ import Bidon
 import VungleAdsSDK
 
 
-final class VungleAdViewContainer: UIView, AdViewContainer {
-    var isAdaptive: Bool { false }
-}
-
-
-final class VungleAdViewDemandProvider: VungleBaseDemandProvider<VungleBanner> {
+final class VungleAdViewDemandProvider: VungleBaseDemandProvider<VungleBannerView> {
     weak var adViewDelegate: DemandProviderAdViewDelegate?
     weak var rootViewController: UIViewController?
 
-    let adSize: BannerSize
+    let adSize: VungleAdSize
+    private weak var banner: VungleBannerView?
+
+    private var hasAdLoaded = false
+
 
     init(context: AdViewContext) {
         self.rootViewController = context.rootViewController
-        self.adSize = BannerSize(format: context.format)
+        self.adSize = context.format.vungleAdSize
 
         super.init()
     }
 
-    override func adObject(placement: String) -> VungleBanner {
-        let banner = VungleBanner(
+    override func adObject(placement: String) -> VungleBannerView {
+        let banner = VungleBannerView(
             placementId: placement,
-            size: adSize
+            vungleAdSize: adSize
         )
         banner.delegate = self
+        self.banner = banner
         return banner
     }
 }
 
 
 extension VungleAdViewDemandProvider: AdViewDemandProvider {
-    func container(for ad: VungleDemandAd<VungleBanner>) -> AdViewContainer? {
-        let rect = CGRect(origin: .zero, size: adSize.cgSize)
-        let container = VungleAdViewContainer(frame: rect)
-        ad.adObject.present(on: container)
-        return container
+    func container(for ad: VungleDemandAd<VungleBannerView>) -> AdViewContainer? {
+        return banner
     }
 
-    func didTrackImpression(for ad: VungleDemandAd<VungleAdsSDK.VungleBanner>) {}
+    func didTrackImpression(for ad: VungleDemandAd<VungleAdsSDK.VungleBannerView>) {}
 }
 
-extension VungleAdViewDemandProvider: VungleBannerDelegate {
-    func bannerAdDidLoad(_ banner: VungleBanner) {
+extension VungleAdViewDemandProvider: VungleBannerViewDelegate {
+    func bannerAdDidLoad(_ bannerView: VungleAdsSDK.VungleBannerView) {
         guard demandAd.adObject === banner else { return }
 
         response?(.success(demandAd))
         response = nil
+
+        hasAdLoaded = true
     }
 
-    func bannerAdDidFailToLoad(_ banner: VungleBanner, withError: NSError) {
+    func bannerAdDidFail(_ bannerView: VungleAdsSDK.VungleBannerView, withError: NSError) {
         guard demandAd.adObject === banner else { return }
 
-        response?(.failure(.noFill(withError.localizedDescription)))
+        response?(.failure(MediationError(error: withError)))
         response = nil
+
     }
 
-    func bannerAdDidFailToPresent(_ banner: VungleBanner, withError: NSError) {
+    func bannerAdDidClose(_ bannerView: VungleAdsSDK.VungleBannerView) {
         guard demandAd.adObject === banner else { return }
 
-        delegate?.provider(
-            self,
-            didFailToDisplayAd: demandAd,
-            error: .generic(error: withError)
-        )
+        delegate?.providerDidHide(self)
     }
 
-    func bannerAdDidTrackImpression(_ banner: VungleBanner) {
+    func bannerAdDidTrackImpression(_ bannerView: VungleAdsSDK.VungleBannerView) {
         guard demandAd.adObject === banner else { return }
 
         revenueDelegate?.provider(self, didLogImpression: demandAd)
     }
 
-    func bannerAdWillLeaveApplication(_ banner: VungleBanner) {}
-
-    func bannerAdDidClick(_ banner: VungleBanner) {
+    func bannerAdDidClick(_ bannerView: VungleAdsSDK.VungleBannerView) {
         guard demandAd.adObject === banner else { return }
 
         delegate?.providerDidClick(self)
     }
+}
 
-    func bannerAdDidClose(_ banner: VungleBanner) {
-        guard demandAd.adObject === banner else { return }
 
-        delegate?.providerDidHide(self)
+extension VungleBannerView: @retroactive AdViewContainer {
+    public var isAdaptive: Bool { false }
+}
+
+
+extension VungleBannerView: VungleLoadableAd {}
+
+
+extension Bidon.BannerFormat {
+    var vungleAdSize: VungleAdSize {
+        switch self {
+        case .banner, .adaptive:
+            return .VungleAdSizeBannerRegular
+        case .leaderboard:
+            return .VungleAdSizeLeaderboard
+        case .mrec:
+            return .VungleAdSizeMREC
+        @unknown default:
+            return .VungleAdSizeBannerRegular
+        }
     }
 }
 
 
-extension BannerSize {
-    init(format: Bidon.BannerFormat) {
-        switch format {
-        case .banner, .adaptive:
-            self = .regular
-        case .leaderboard:
-            self = .leaderboard
-        case .mrec:
-            self = .mrec
-        }
-    }
-
-    var cgSize: CGSize {
-        switch self {
-        case .mrec:
-            return CGSize(width: 300, height: 250)
-        case .leaderboard:
-            return CGSize(width: 728, height: 90)
-        case .regular:
-            return CGSize(width: 320, height: 50)
-        case .short:
-            return CGSize(width: 300, height: 50)
-        @unknown default:
-            return .zero
+extension MediationError {
+    init(error: NSError) {
+        switch error.code {
+        case 101, 102, 103, 104, 105, 106, 110, 111, 118, 119, 122, 124, 125, 134, 135, 138, 20001:
+            self = .networkError
+        case 208, 209, 214:
+            self = .noBid(error.localizedDescription)
+        case 213:
+            self = .incorrectAdUnitId
+        case 10001:
+            self = .noFill(error.localizedDescription)
+        default:
+            self = .unspecifiedException("Code: \(error.code), \(error.localizedDescription)")
         }
     }
 }