+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # develop
 
+# Release 0.9.0
+
 ## New features
+- BDN-987 VungleSDK Banner API - updated
+- APDM-1396 Update BCA MAX callbacks
+- BDN-966 Updated Bidmachine to 3.3.0. Added BM placements logic
+- APDM-1411 Update IronSource network
 
 # Release 0.8.1